refactor(tests): 完成测试代码整合与 Clippy 修复
- 修复所有 Clippy 警告 (uninlined_format_args, redundant_closure, doc_markdown, ptr_arg 等)
- 将 root-level 测试文件整合到模块化结构
- 删除零散的 root-level 测试文件 (http.rs, proxy_commands.rs 等)
- 新增模块化测试文件 (git/adaptive_tls.rs, git/submodule.rs, proxy/manager_*.rs 等)
- 新增 workspace/ 模块并整合所有工作区测试
- 修复 Cargo.toml bench 配置 (添加 required-features)
- 所有测试通过，0 警告 0 错误

+++ src-tauri/src/core/ip_pool/sampling.rs
@@ -65,7 +65,7 @@ pub(super) async fn ensure_sampled(
 
         if let Some(waiter) = waiter {
             let timeout_ms = pool.config.runtime.singleflight_timeout_ms.max(100);
-            if timeout(Duration::from_millis(timeout_ms as u64), waiter.notified())
+            if timeout(Duration::from_millis(timeout_ms), waiter.notified())
                 .await
                 .is_err()
             {
